# apps/api/src/api-docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tickets/{ticketId}/messages": {
      "get": {
        "operationId": "MessagesController_findByTicket",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "Ticket ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all messages in a ticket",
        "tags": [
          "Messages"
        ]
      },
      "post": {
        "operationId": "MessagesController_create",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "Ticket ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        },
        "summary": "Create a message in a ticket",
        "tags": [
          "Messages"
        ]
      }
    },
    "/api/tickets/{ticketId}/messages/{messageId}": {
      "delete": {
        "operationId": "MessagesController_delete",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "description": "Message ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "Ticket ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a message",
        "tags": [
          "Messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "TICOAI API",
    "description": "AI-Powered Customer Support Ticket System API",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "schemas": {
...
      "Message": {
        "type": "object",
        "properties": {}
      }
    }
  }
}