# packages/api-sdk/src/generated/schemas/message.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * TICOAI API
 * AI-Powered Customer Support Ticket System API
 * OpenAPI spec version: 2.0.0
 */

export type Message = { [key: string]: unknown };